test: move tests from src to lib
